+++ changelog.md
@@ -1,2 +1,2 @@
 ### Ore Growth 1.1.7
-- Initial release of Ore Growth for Minecraft 1.21.9 & 1.21.10
+- Initial release of Ore Growth for Minecraft 1.21.11